requirements.txt
feedparser
spotipy
pyyaml
beautifulsoup4
requests
requests-cache
anthropic
google-genai
python-dotenv
pydantic-settings
fastapi
uvicorn
aiofiles
twilio
apscheduler
groq
pdfplumber
pytest
responses
ruff
mypy
